preparing for wallet-connect integration
